+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### New
 
 ### Changes
+- adding support for module-type spec on init of new module `seedfarmer init module -mt cdkv2`
 
 ### Fixes